+++ backend-spring/src/main/java/com/example/factoryguard/application/dto/operation/SystemStatusResult.java
@@ -3,16 +3,15 @@ package com.example.factoryguard.application.dto.operation;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
 @Builder
 public class SystemStatusResult {
 
-    private final BigDecimal cpuUsage;
-    private final BigDecimal memoryUsage;
-    private final BigDecimal diskUsage;
+    private final Double cpuUsage;
+    private final Double memoryUsage;
+    private final Double diskUsage;
     private final Integer responseTimeMs;
     private final LocalDateTime snapshotAt;
 }
