+++ backend-spring/src/main/java/com/example/factoryguard/application/dto/result/AnomalyRegionResponse.java
@@ -3,17 +3,15 @@ package com.example.factoryguard.application.dto.result;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Builder
 public class AnomalyRegionResponse {
 
     private final Long regionId;
     private final String labelCode;
-    private final BigDecimal bboxX;
-    private final BigDecimal bboxY;
-    private final BigDecimal bboxW;
-    private final BigDecimal bboxH;
-    private final BigDecimal score;
+    private final Double bboxX;
+    private final Double bboxY;
+    private final Double bboxW;
+    private final Double bboxH;
+    private final Double score;
 }
