+++ backend-spring/src/main/java/com/example/factoryguard/application/dto/result/AnomalyRegionResult.java
@@ -4,8 +4,6 @@ import com.example.factoryguard.domain.result.vo.AnomalyRegionLabel;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Builder
 public class AnomalyRegionResult {
@@ -13,9 +11,9 @@ public class AnomalyRegionResult {
     private final Long regionId;
     private final Long imageId;
     private final AnomalyRegionLabel labelCode;
-    private final BigDecimal bboxX;
-    private final BigDecimal bboxY;
-    private final BigDecimal bboxW;
-    private final BigDecimal bboxH;
-    private final BigDecimal score;
+    private final Double bboxX;
+    private final Double bboxY;
+    private final Double bboxW;
+    private final Double bboxH;
+    private final Double score;
 }
