+++ backend-spring/src/main/java/com/example/factoryguard/application/dto/result/ResultListSummaryResponse.java
@@ -3,8 +3,6 @@ package com.example.factoryguard.application.dto.result;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Builder
 public class ResultListSummaryResponse {
@@ -13,5 +11,5 @@ public class ResultListSummaryResponse {
     private final long normalCount;
     private final long defectCount;
     private final long retestCount;
-    private final BigDecimal avgScore;
+    private final Double avgScore;
 }
