+++ backend-spring/src/main/java/com/example/factoryguard/domain/operation/model/SystemStatusSnapshot.java
@@ -3,7 +3,6 @@ package com.example.factoryguard.domain.operation.model;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -11,9 +10,9 @@ import java.time.LocalDateTime;
 public class SystemStatusSnapshot {
 
     private final Long snapshotId;
-    private final BigDecimal cpuUsage;
-    private final BigDecimal memoryUsage;
-    private final BigDecimal diskUsage;
+    private final Double cpuUsage;
+    private final Double memoryUsage;
+    private final Double diskUsage;
     private final Integer responseTimeMs;
     private final LocalDateTime createdAt;
 }
