+++ backend-spring/src/main/java/com/example/factoryguard/domain/result/model/AnomalyRegion.java
@@ -4,7 +4,6 @@ import com.example.factoryguard.domain.result.vo.AnomalyRegionLabel;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -14,10 +13,10 @@ public class AnomalyRegion {
     private final Long regionId;
     private final Long imageId;
     private final AnomalyRegionLabel labelCode;
-    private final BigDecimal bboxX;
-    private final BigDecimal bboxY;
-    private final BigDecimal bboxW;
-    private final BigDecimal bboxH;
-    private final BigDecimal score;
+    private final Double bboxX;
+    private final Double bboxY;
+    private final Double bboxW;
+    private final Double bboxH;
+    private final Double score;
     private final LocalDateTime createdAt;
 }
